+++ build64/Makefile
@@ -46,6 +46,7 @@ SYS_OBJS = system/acpi.o \
            system/serial.o \
            system/smbios.o \
            system/smbus.o \
+           system/spd.o \
            system/smp.o \
            system/temperature.o \
            system/timers.o \
